handle ERR_NAMEINUSE
this changes a random character of the nickname if it recives
ERR_NAMEINUSE (433), to prevent clashes while also keeping the same
length. this will result in an infinite loop of trying nicknames if
there are none available with the same length.

FIXME: it lazily tries only characters known to be allowed anywhere in a
nickname. this means it will never introduce a number into a nickname.